Create the seed of an ImContext class
The documentation for the class explains why it exists in more details.
Anyhow, the idea is that we want to be able to support chorded input for
characters, hence we need to use GtkImContext, which has no C++ wrapper.
This is eventually going to become a more or less complete wrapper.
Right now, it just eats characters, and then prints out what the result
will be. The next step would be to create a signal producing a ustring
that we can actually subscribe to.